+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 MikoPBX module for callerid resolution via suisse directory api of https://tel.search.ch
 
+## [x.x.x] not yet released
+### Added
+- drop anonymous calls or call from category callcenter on demand
+
 ## [1.1.0] - 2026-09-21
 ### Fixed
 - some suggestions from MikoPBX which should be essentially included on a mikopbx module